[Refactor] Extract `ICalendar.vAlarmToMin` to `AlarmTriggerCalculator`
