+++ lib/src/main/kotlin/at/bitfire/synctools/mapping/tasks/DmfsTaskBuilder.kt
@@ -6,7 +6,6 @@
 
 package at.bitfire.synctools.mapping.tasks
 
-import at.bitfire.ical4android.ICalendar
 import at.bitfire.ical4android.Task
 import at.bitfire.ical4android.UnknownProperty
 import at.bitfire.synctools.icalendar.DatePropertyTzMapper.normalizedDate
@@ -16,6 +15,7 @@ import at.bitfire.synctools.storage.tasks.DmfsTask.Companion.COLUMN_FLAGS
 import at.bitfire.synctools.storage.tasks.DmfsTask.Companion.UNKNOWN_PROPERTY_DATA
 import at.bitfire.synctools.storage.tasks.DmfsTaskList
 import at.bitfire.synctools.storage.tasks.TasksBatchOperation
+import at.bitfire.synctools.util.AlarmTriggerCalculator
 import at.bitfire.synctools.util.AndroidTimeUtils
 import at.bitfire.synctools.util.AndroidTimeUtils.toTimestamp
 import net.fortuna.ical4j.model.Parameter
@@ -208,7 +208,7 @@ class DmfsTaskBuilder(
 
     private fun insertAlarms(batch: TasksBatchOperation, idxTask: Int?) {
         for (alarm in task.alarms) {
-            val (alarmRef, minutes) = ICalendar.vAlarmToMin(
+            val (alarmRef, minutes) = AlarmTriggerCalculator.vAlarmToMin(
                 alarm = alarm,
                 refStart = task.dtStart,
                 refEnd = task.due,
